proyectos/servicios/views.py
from rest_framework import viewsets
from rest_framework import filters
from .models import Usuario, Servicio, Solicitud, Calificacion
from .serializers import UsuarioSerializer, ServicioSerializer, SolicitudSerializer, CalificacionSerializer

class UsuarioViewSet(viewsets.ModelViewSet):
    queryset = Usuario.objects.all()
    serializer_class = UsuarioSerializer

class ServicioViewSet(viewsets.ModelViewSet):
    queryset = Servicio.objects.all()
    serializer_class = ServicioSerializer
    filter_backends = [filters.SearchFilter]
    search_fields = ['titulo', 'categoria']

class SolicitudViewSet(viewsets.ModelViewSet):
    queryset = Solicitud.objects.all()
    serializer_class = SolicitudSerializer

class CalificacionViewSet(viewsets.ModelViewSet):
    queryset = Calificacion.objects.all()
    serializer_class = CalificacionSerializer
